Update setup.py to work for both python 2 and 3.
Using distribute's ability to run 2 to 3 for you automatically, the
library will now build correctly for python 3.  In order for the
versioning to work correctly the version number had to be pulled out of
the __init__.py file since that file would not import correctly on
python 3 systems and so setup would not run at all.