Use the IUGG mean earth radius for EARTH_RADIUS.
The previous value was the "quadratic mean radius" which is defined here

  http://math.wikia.com/wiki/Ellipsoidal_quadratic_mean_radius

and was, for a time, touted on Wikipedia.  Here there's a flaw in the
derivation of this quantity -- the way the azimuths was sampled was
wrong; see

  https://en.wikipedia.org/wiki/Talk:Great-circle_distance/Archive_1

The IUGG value minimizes the RMS difference between the spherical and
ellipsoidal distances.