Docs: replace toctree with links
Toctree cannot reference headings of the same page (at least
in a straightforward way), and without them the ToC of
the generated document is poorly browsable w/o these headings.

I'm not willing to split the docs and move them to (a) separate
file(s) right now, so let's return back the previous (1.11.0) edition
(except that 1xx Changelog is also a link now, though it was
included before).