Tests: merge nominatim.py and nominatim_base.py
The reason why I've split them in the first place is that I thought
that importing `BaseNominatimTestCase` from `nominatim.py` in
`pickpoint.py` would also execute `NominatimTestCase`, which would
effectively run the same tests twice (both from `nominatim.py`
and `pickpoint.py`).

Indeed, when `NominatimTestCase` is imported in `pickpoint.py`,
the tests are run twice (b/c the `NominatimTestCase` reference exists
in both modules). But there's no point in importing it! All we need
in `pickpoint.py` is `BaseNominatimTestCase`.

Having a single module for nominatim tests looks much better to me,
as it avoids confusion and is cleaner in terms of modules structure.

Don't know what I was thinking yesterday, but this commit fixes that
foolish mistake.