Docs: add toctree, make sidebar nav a bit better
I wanted to have all geocoders listed in the sidebar toctree, because
viewing docs for a specific geocoder is quite a common use-case.
Apparently without introducing level 3 headers that's impossible [1].

sphinx-rtd-theme for some reason cannot render third level of headings
correctly in the sidebar without a `toctree` present in the doc:
they appear at the second level along with the other level two headers.
Hence the need to have `toctree` in the doc. It could've been hidden,
but I figured it might be helpful to have it visible in the doc.
Will see how it goes.

The toctree references the same document, which causes a warning
regarding a circular reference, yet it's the only option I could
find to have the third level of headings be correctly rendered in
the sidebar [2].

[1]: https://stackoverflow.com/a/18002745
[2]: https://github.com/sphinx-doc/sphinx/issues/2103