Geocodefarm tests: add GEOCODEFARM_SKIP env
The service seems to experience some issues currently. It always
returns empty responses. The status panel https://geocode.farm/serverstatus.php
indicates that there's an open issue with exactly that (which seems
to be open for ~5 days now *raised eyebrows*).

Work-around that by introducing an env variable, which, when set,
marks the testcase as skipped.