Deprecate `format_string`, add `Specifying Parameters Once` doc section
The main reason for that is that it is too easy to forget about
`format_string` when adding a new geocoder. This complicates maintenance.

Another reason -- there're other methods and params that might need
a similar templating. Python already contains great tools for this
problem (see the new doc section), so `format_string` seems to be
redundant.