Allowing different mapbox endpoints.
The mapbox api provides different endpoints that are
differentiated by the url path. Adding an endpoint
parameter that is used to template the api path and
providing a default parameter to preserve the current
behaviour. Doing so allows us to create a mapbox
geocoder and call the mapbox.places-permanent endpoint.