Fix Github actions build badge
See the linked GitHub issue from the previous version of the badge: https://github.com/badges/shields/issues/8671

In summary, these badges were changed in a breaking fashion such that the badges were just linking to the github issue rather than showing the data. Updating the url resolves this.

Also updated the link added to the badge image, so that it's restricted to the specific action and the branch in the way the badge is.